requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.8
django-import-export==2.0
django-polymorphic==2.1.2
https://github.com/darklow/django-suit/tarball/v2
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
jdcal==1.4.1
MarkupPy==1.14
numpy==1.17.4
odfpy==1.4.0
openpyxl==2.6.4
pandas==0.24.2
Pillow==6.2.1
psycopg2==2.8.4
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
reportlab==3.5.32
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
webencodings==0.5.1
whitenoise==4.1.4
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
django-heroku==0.3.1
